refactor: rename `parquet_file::{utils => test_utils}`
