+++ parquet_file/src/lib.rs
@@ -13,6 +13,6 @@ pub mod cleanup;
 pub mod metadata;
 pub mod storage;
 pub mod table;
-pub mod utils;
+pub mod test_utils;
 
 mod storage_testing;
